requirements.txt
biopython==1.78
matplotlib==3.5.3
numpy==1.21.6
pandas==1.3.5
